+++ test_settings.py
@@ -28,6 +28,7 @@ DATABASES = {
 
 INSTALLED_APPS = (
     'config_models',
+    'django.contrib.admin',
     'django.contrib.auth',
     'django.contrib.contenttypes',
     'django.contrib.sessions',